Transition kernel to x86-64 long mode (4-level paging)
Boot now enables PAE + EFER.LME + CR0.PG before jumping to
CODE64 (0x18), putting the kernel in true 64-bit long mode.
User programs continue to run in 32-bit compatibility mode
(CS=0x20, DPL=3) so all existing ELF32 binaries work unchanged.

Key changes:
- boot.asm: set PML4/PDPT/PD, CR4.PAE, EFER.LME, CR0.PG, then
  far-jump to CODE64; fix RELOC_DELTA on long_mode_64 target
- gdt.asm: add CODE64 descriptor (L=1) at 0x18; extend TSS to
  16-byte 64-bit descriptor; add USER_CODE/DATA at 0x20/0x28
- kernel.ld: add .rdata, _text_end alias, discard .pdata/.xdata
- kernel_gfx.ld: new — 32-bit linker script for gfx demo
- build.bat: KFLAGS=-m64 -mno-red-zone, elf64, i386pep linker
- run.bat: switch to qemu-system-x86_64 (i386 lacks EFER.LME)
- idt.asm: full 64-bit IDT (16-byte entries, 15-GPR save, iretq)
- syscalls.asm: read AH via [rsp+1] to avoid REX+high-byte conflict
- tasking.c/h: 64-bit RSP, page_directory, schedule() signature
- paging.c/h: 4-level tables (PML4→PDPT→PD→PT), NX support
- tss.c/h: RSP0 at offset 4 (8 bytes), 104-byte TSS64
- usermode.asm: iretq with CS=USER_CODE32|3 for compat mode entry
- heap.c: pushfl/popl → pushfq/popq for 64-bit flags save
- spinlock.h: same pushfq/popq fix for spin_save_and_disable_irq
- selftest.c: cr0/cr4 vars widened to unsigned long long

+++ src/kernel/selftest.c
@@ -24,7 +24,7 @@ static int buf_eq(unsigned char* a, unsigned char* b, unsigned int len) {
 }
 
 static int test_paging() {
-    unsigned int cr0, cr4;
+    unsigned long long cr0, cr4;
     __asm__ volatile("mov %%cr0, %0" : "=r"(cr0));
     __asm__ volatile("mov %%cr4, %0" : "=r"(cr4));
